refactor(types): move swarm common types into types module
